graph drawing: Reimplement the network simplex using recursive DFS.
This makes it more similar to the original GansnerKNV implementation and
to the paper. The recursive functions are less error-prone and finally
work! In future commits I will try to implement 1-2 generic
pre-/post-order DFS traversals in order to bring iterative DFS back
without introducing the same weird bugs again that are really hard to
trace due to the amount of DFS boilerplate code needed.