graph drawing: Rework loop and multiedge handling.
The Edge class no longer checks whether a node is already part of the
edge. NOTE: This can lead to problems in existing algorithms, so we need
to verify that those still work.

The result is that loops contain the same node twice. Algorithms can
detect loops by testing whether tail and head are the same.

The "modular layered" algorithm drops loops prior to the cycle removal
step and adds them back to the graph before routing the edges. It also
merges multiedges and restores the original edges prior to the same
steps as loops.

Again, other algorithms might have to be adjusted to work again (if they
are broken).